add car loan. add admin dashboard. fix bugs

+++ src/main/java/com/perfiosbank/closeaccount/CloseAccountService.java
@@ -1,6 +1,7 @@
 package com.perfiosbank.closeaccount;
 
 import com.perfiosbank.exceptions.AccountNotFoundException;
+import com.perfiosbank.exceptions.ActiveLoansFoundException;
 import com.perfiosbank.exceptions.AuthenticationFailedException;
 import com.perfiosbank.model.User;
 import com.perfiosbank.utils.AccountUtils;
@@ -8,7 +9,8 @@ import com.perfiosbank.utils.AuthenticationUtils;
 
 public class CloseAccountService {
     public void closeAccount(User userInSession, User enteredDetails) 
-    		throws AuthenticationFailedException, AccountNotFoundException, Exception {
+    		throws AuthenticationFailedException, AccountNotFoundException, 
+    		ActiveLoansFoundException, Exception {
         String msg;
 
         if (AuthenticationUtils.isUserNotAuthenticated(userInSession, enteredDetails)) {
@@ -21,6 +23,11 @@ public class CloseAccountService {
             throw new AccountNotFoundException(msg);
         }
 
+        if (CloseAccountDao.getNumberOfApprovedLoans(userInSession.getUsername()) != 0) {
+        	msg = "You cannot close your account when you have active loans to clear!";
+        	throw new ActiveLoansFoundException(msg);
+        }
+        
         if (CloseAccountDao.removeUser(userInSession) != 1) {
         	throw new Exception();
         }
